This commit adds exception handling and logging for all the services. And removes a bug where a user can make multiple pending book requests on the same book.

+++ src/main/java/com/zolobooky/booky/images/BlobService.java
@@ -3,6 +3,9 @@ package com.zolobooky.booky.images;
 import com.azure.storage.blob.BlobClient;
 import com.azure.storage.blob.BlobContainerClient;
 
+import com.azure.storage.blob.models.BlobStorageException;
+import com.zolobooky.booky.images.ImageExceptions.ImageNotFoundExceptions;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
@@ -10,6 +13,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 
+@Slf4j
 @Component
 public class BlobService {
 
@@ -35,8 +39,15 @@ public class BlobService {
 
 		ByteArrayOutputStream outStream = new ByteArrayOutputStream();
 
-		blob.downloadStream(outStream);
+		try {
+			blob.downloadStream(outStream);
 
+		}
+		catch (BlobStorageException be) {
+			log.info(String.format("Image with name: %s not found to download", fileName));
+			throw new ImageNotFoundExceptions(String.format("Image with name: %s not found to download", fileName));
+		}
+		log.info(String.format("Image with name: %s downloaded", fileName));
 		return outStream.toByteArray();
 	}
 
@@ -44,6 +55,12 @@ public class BlobService {
 
 		BlobClient blob = blobContainerClient.getBlobClient(fileName);
 
+		if (!blob.exists()) {
+			log.info(String.format("Image with name: %s not found to delete", fileName));
+			throw new ImageNotFoundExceptions(String.format("Image with name: %s not found to delete", fileName));
+		}
+		;
+		log.info(String.format("Image with name: %s deleted", fileName));
 		return blob.deleteIfExists();
 	}
